fix: Update path for releases (#25)
## Which problem is this PR solving?
The path `processor` was conflicting with a folder in the upstream
collector repo when installing this as a module. This PR moves to a more
specific name `symbolicatorprocessor` to avoid that conflict. An
alternative would be to consider moving everything to the top level
directory but this leaves things open if we want to have other collector
components here in the future.

## Short description of the changes
- Changed `processor` directory to `symbolicatorprocessor`

## How to verify that this has the expected result
- Tests pass
- Collector builds