feat: make the source map symbolicator use the debug uuid attribute, if present (#113)
## Which problem is this PR solving?

The source map processor expects file names to have a unique UUID in
them. But JS files created from frameworks like React Native have fixed
names like "index.bundle". So, this PR adds a new feature, where, if the
trace has a debug UUID specified in the resource attributes, it will be
used as a directory to look inside for the source file.

## Short description of the changes

1. If the UUID is present, put it in the path.
2. Fix "native" frames by having the symbolicator pass through frames
with no url, instead of returning an error.

## How to verify that this has the expected result

Unit tests are added. I also manually tested by running the collector
locally and sending it stacktraces with the UUID attached and verifying
it symbolicated them correctly.

---

- [X] CHANGELOG is updated
N/A ~- [ ] README is updated with documentation~