fix(sourcemap-processor): add handling for anonymous urls (#141)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Short description of the changes

The processor was throwing symbolication errors on frames with anonymous
urls because the symbolication logic didn't know how to handle this edge
case. Adding specail handling for `<anonymous` URL frames. We are
skipping symbolication and printing out the frame as-is.

## How to verify that this has the expected result
Unit tests pass! Tested locally.

---

- [X] CHANGELOG is updated
- [ ] README is updated with documentation